feat: 支持upsert的sql编译

+++ lib/core/compiler_remove.go
@@ -5,6 +5,10 @@ import (
 )
 
 func (my *compilerContext) renderRemove(id, pid int, f *ast.Field) {
+	remove := f.Arguments.ForName(REMOVE)
+	if remove == nil || remove.Value.Raw != "true" {
+		return
+	}
 	table, _ := my.meta.TableName(f.Definition.Type.Name(), false)
 	my.Quoted(table)
 	my.Space(`AS (DELETE FROM`)
